Got SetUpvalue to work somehow
Still not sure if everything works exactly as it should,
need to formulate more tests to weed out any strange corner cases
not mentioned in http://craftinginterpreters.com either because
they "just work" in the clox implementation or because the author
wasn't mad enough to come up with them.

Also need to do some serious cleanup and correct the comments which
are probably inaccurate in many places because I've misunderstood
exactly how things work even in my own implementation.

If memory management in C is hard and implementing upvalues is
hard squared, then translating C code that implements upvalues to
Rust without breaking anything (while trying to learn Rust!)
is something like hard cubed.

+++ src/lox/callframe.rs
@@ -13,13 +13,13 @@ use super::closure::Closure;
 pub struct CallFrame {
     closure_value:	Value,
     ip: 		u32,
-    stack_bottom:	u32,
+    stack_bottom:	usize,
 }
 
 
 //#[allow(dead_code)]
 impl CallFrame {
-    pub fn new(closure: Value, stack_bottom: u32) -> CallFrame {
+    pub fn new(closure: Value, stack_bottom: usize) -> CallFrame {
         if !closure.is_closure() {
             panic!("{} is not a Closure", closure);
         }
@@ -93,7 +93,7 @@ impl CallFrame {
         self.ip = ip;
     }
     
-    pub fn stack_bottom(&self) -> u32 {
+    pub fn stack_bottom(&self) -> usize {
         return self.stack_bottom;
     }
 }
